JENKINS-26097: Adjust label expression auto-completion and validation.
This moves label expression auto-completion and validation from AbstractProject
to LabelExpression.

Auto-Completion:
- AbstractProject.AutoCompleteSeerder was moved to LabelExpression
  - test class correspondingly moved to the relevant package
- added static LabelExpression.autoComplete() method
  - does what the non-static AbstractProject.doAutoCompleteLabel() used to do; that now calls the new static method

Validation:
- added LabelExpression.LabelValidator, which is like AbstractProject.LabelValidator except it takes a Job instead of an AbstractProject
  - AbstractProject.LabelValidator is now deprecated
- added static LabelExpression.validate(String, Job) method, which replaces AbstractProject.validateLabelExpression(String, AbstractProject)
  - the latter is marked as deprecated and forwards to the former
- LabelExpressionTest.formValidation() uses the new static method
